+++ plugin/src/test/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaKotlinRouteCollectorTest.kt
@@ -1,0 +1,43 @@
+package com.linecorp.intellij.plugins.armeria.explorer
+
+import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
+
+class ArmeriaKotlinRouteCollectorTest : LightJavaCodeInsightFixtureTestCase() {
+    override fun setUp() {
+        super.setUp()
+        registerArmeriaStubs()
+    }
+
+    fun testCollectAnnotatedRouteFromKotlinLightMethod() {
+        myFixture.configureByText(
+            "HelloService.kt",
+            """
+            package example
+
+            import com.linecorp.armeria.server.annotation.Get
+
+            class HelloService {
+                @Get("/hello")
+                fun hello(): String = "hello"
+            }
+            """.trimIndent(),
+        )
+
+        val routes = ArmeriaRouteCollector.collect(project)
+        val kotlinRoute = routes.firstOrNull { it.path == "/hello" && it.httpMethod == "GET" }
+        assertNotNull("Expected Kotlin annotated route to be discovered", kotlinRoute)
+    }
+
+    private fun registerArmeriaStubs() {
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.server.annotation;
+
+            public @interface Get {
+                String value() default "";
+                String path() default "";
+            }
+            """.trimIndent(),
+        )
+    }
+}